Move ManageColumnContext to ManageColumns
