+++ src/components/DataTableControls/DataTableControls.test.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import DataTableControls from ".";
-import ManageColumnsContext from '../DatasetTableTab/ManageColumnsContext';
+import ManageColumnsContext from '../ManageColumns/ManageColumnsContext';
 import DataTableContext from "../../templates/Dataset/DataTableContext";
 import * as resource from "../../tests/fixtures/resource.json";
 import * as distribution from "../../tests/fixtures/distribution.json";
