+++ src/components/DatasetTableTab/DataTableStateWrapper.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { createContext, useContext, useState } from "react";
 import DatasetTable from ".";
 import DataTableContext from "../../templates/Dataset/DataTableContext";
-import ManageColumnsContext from './ManageColumnsContext';
+import ManageColumnsContext from '../ManageColumns/ManageColumnsContext';
 
 const DataTableStateWrapper = () => {
   const { id, datasetTableControls } = useContext(DataTableContext);
