+++ src/components/DatasetTableTab/index.tsx
@@ -8,7 +8,7 @@ import DataTableHeader from '../DatatableHeader';
 import QueryBuilder from '../QueryBuilder';
 import { DistributionType, ColumnType, ResourceType } from '../../types/dataset';
 import DataTableContext from "../../templates/Dataset/DataTableContext";
-import ManageColumnsContext from '../DatasetTableTab/ManageColumnsContext';
+import ManageColumnsContext from '../ManageColumns/ManageColumnsContext';
 
 export function prepareColumns(columns : any, schema : any) {
   return columns.map((column : any) => ({
