+++ src/templates/FilteredResource/FilteredResourceBody.jsx
@@ -13,7 +13,7 @@ import TransformedDate from '../../components/TransformedDate';
 import FilteredResourceDescription from './FilteredResourceDescription';
 import 'swagger-ui-react/swagger-ui.css';
 import DataTableContext from '../Dataset/DataTableContext';
-import ManageColumnsContext from '../../components/DatasetTableTab/ManageColumnsContext';
+import ManageColumnsContext from '../../components/ManageColumns/ManageColumnsContext';
 
 const FilteredResourceBody = ({
   id,
